PageSwipe issue fixed.
