+++ lib/blocs/article_list/tech_bloc.dart
@@ -1,14 +1,14 @@
 import 'package:bloc/bloc.dart';
-import 'package:meta/meta.dart';
-import 'package:newzzflutter/blocs/article_list/article_list_event.dart';
-import 'package:newzzflutter/blocs/article_list/article_list_state.dart';
 import 'package:newzzflutter/domain/model/article_response_model.dart';
 import 'package:newzzflutter/domain/repositories/article_repo.dart';
+import 'package:meta/meta.dart';
 
-class ArticleBloc extends Bloc<ArticleEvent, ArticleState> {
+import 'article_list_event.dart';
+import 'article_list_state.dart';
+class TechnologyBloc extends Bloc<ArticleEvent, ArticleState> {
   final ArticleRepo articleRepo;
 
-  ArticleBloc({@required this.articleRepo});
+  TechnologyBloc({@required this.articleRepo});
 
   @override
   ArticleState get initialState => ArticleLoadingState();
