+++ lib/ui/common_widgets/article_row_widget.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_custom_tabs/flutter_custom_tabs.dart';
 import 'package:newzzflutter/constants/dimens.dart';
 import 'package:newzzflutter/domain/model/article_response_model.dart';
 
@@ -81,34 +80,5 @@ class ArticleWidget extends StatelessWidget {
         ),
       ),
     );
-  }
-}
-
-void _launchURL(BuildContext context, String url) async {
-  try {
-    await launch(
-      '$url',
-      option: new CustomTabsOption(
-        toolbarColor: Theme.of(context).primaryColor,
-        enableDefaultShare: true,
-        enableUrlBarHiding: true,
-        showPageTitle: true,
-        animation: new CustomTabsAnimation(
-          startEnter: 'slide_up',
-          startExit: 'android:anim/fade_out',
-          endEnter: 'android:anim/fade_in',
-          endExit: 'slide_down',
-        ),
-        extraCustomTabs: <String>[
-          // ref. https://play.google.com/store/apps/details?id=org.mozilla.firefox
-          'org.mozilla.firefox',
-          // ref. https://play.google.com/store/apps/details?id=com.microsoft.emmx
-          'com.microsoft.emmx',
-        ],
-      ),
-    );
-  } catch (e) {
-    // An exception is thrown if browser app is not installed on Android device.
-    debugPrint(e.toString());
   }
 }
